Add options to prevent compilation error on Java 24
Newer versions of Java have added a limit to the size.
This causes a compilation error when compiling the Java code,
since some of the Tycho-related XML files are very large.
Increasing this limit through an option allows the Java code to be
compiled properly on Java 24 and newer.

See https://github.com/eclipse-jdtls/eclipse.jdt.ls/issues/3465#issuecomment-2959641987

Signed-off-by: David Thompson <davthomp@redhat.com>